Add php-mode-maybe for auto-mode-alist
The new function `php-mode-maybe` is wrapper for auto-mode-alist that
invokes another function that is not php-mode under certain conditions.

One of them is the Bladetemplate.  This file generally has .blade.php
extension, but it can not be edited in php-mode.  Fortunately web-mode
supports Blade templates.  When web-mode is not installed in Emacs,
simply announce warn and notify the user that php-mode is not suitable
for editing the blade file.