[Fix] Correct the package metadata for the version number
PHP Mode is at version 1.18.3 but the package metadata still had the
previous version number.  This patch corrects that by making the two
equal as they should be.  And we can consider this a proper bug-fix
bump in the version thanks to the work of USAMI Kenta for fixing a
problem with `skeleton/php-current.el`.

GitHub-Issue: #323
GitHub-Issue: #359
Signed-off-by: Eric James Michael Ritz <ericjmritz@yandex.com>

The two GitHub issues above are in reference to the bug fix for
`skeleton/php-current.el`, not the bug fix of correcting the version
number which this patch does.  We reference those issues so that
future developers will be able to easily see what bugs motivated an
increase in the "patch" part of the version number, i.e. going from
v1.18.2 to v1.18.3.