[Docs] Describe the ideal quality of commit messages in the README
Like the previous patch, a40c5e51, this commit slightly updates the
"How to Contribute" section of the README.  We add an additional
suggestion about how contributors should ideally write their commit
messages for all pull-requests.  We make it clear that we will not
absolutely require detailed commit messages for *every* patch, and
sometimes we will even write the commit messages of other developers,
but nonetheless it is important to provide developers with our
guidelines for commit messages.  Those guidelines are important
because they make it much easier for future contributors to understand
the purpose of each change to PHP Mode, and that can help find bugs,
optimally improve functionality, and more.

Signed-off-by: Eric James Michael Ritz <ericjmritz@yandex.com>