Fix Travis CI setup for Emacs snapshots
The current Travis CI setup successfully installs the package
emacs-snapshot when tests against the development version of Emacs are
requested, but does not point the binary "emacs" at the snapshot
version, thus effectively running the test suite against the default
Emacs version (24.3.1).

This change invokes update-alternatives after the installation to make
the binary "emacs" refer to the snapshot version.  Additionally, it
does no longer (unsuccessfully) execute the Emacs Version Manager when
the snapshot setup is requested.