adapt new travis environment & more emacs versions
According to
https://github.com/rejeep/evm/blob/72f90d43fcac7861b620f2b6312e44e0604792e2/README.md
the *-bin packages are going to be deprecated because travis is moving
towards a container infrastructure.

Add two more recent emacs versions, 25.2 and 25.3, to the build
matrix.

Replace the ubuntu git-snapshot package with the one provided by evm,
this also ensures the correct version of emacs is running the tests.