change style initialization to interactive form php-set-style
to be able to select the php-mode code style with the remapped
c-set-style key binding, to have a cleaner code base and reliable
setting of default values over previously set values when changing
the style.

style variables tab-width, fill-column, require-final-newline and
show-trailing-whitespace which are not present in all available
php-mode code styles get their values set from the global default
value on file eval time to the default "php" style.

c-set-style was called twice before this change, now it is called
once, so are the style mode hooks.