derive from prog-mode not from c-mode
Deriving from c-mode results in a few unwanted side effects like the
execution of the c-mode-hook everytime php-mode is enabled.  Deriving
from a major mode makes only sense if the mode is a specialization of
the derived mode (like ng2-mode deriving from typescript mode as it
adds only features and it makes sense that the typescript hook is
being run).  Also there could be some workarounds for c-mode specific
issues which can be removed now (but haven't for the most now)

This commit does not touch the usage of cc-mode stuff like
c-lang-defconst and c-add-language.  Instead of deriving c-mode to
inherit the syntax table c-populate-syntax-table is called directly.

The only workaround removed was the defcustom var
php-do-not-use-semantic-imenu which was a workaround for a broken
imenu when using semantic-mode (as c-mode enables semantic-imenu and
the setting was inherited) which is not necessary anymore.

Github Issue: #407