Fix delayed font-locking issue
Due to a previous refactoring of the font-locking a bug was introduced
where font-locking was delayed. This happened because
c-before-font-lock-functions did not include the function
c-change-expand-fl-region. This commit adds it back to the cc-mode
variable, so font-locking will not be delayed.

c-depropertize-new-text was introduced in current emacs master (will
probably be part of emacs 26?) and c-change-expand-fl-region was
introduced in emacs 25. These will be checked before calling
c-lang-defconst.

Github Issue: #414