set c-before-font-lock-functions to default 't' value to not inherit java's value
this will make the variable more portable through emacs versions

Background:
java has to function in the variable we do not want:
- c-parse-quotes-after-change
- c-restore-<>-properties

as those are the only ones making java mode differ from the default
value we use that instead. The default value (which will be used if
no entry is in c-before-font-lock-functions for the used mode) has
changed from emacs 24 to master several times and may change in the
future. So this solution might spare us a few compability commits in
the future.