Translate README.ja.md
I would like to improve the quality of the whole documents.

This translation is also for Japanese users, but the biggest
motivation is that the current national language of the
maintainer (just me) is Japanese.

He (I) is not good at English, so you can build a logical document by
working on Japanese. Based on that changes, it will be possible to
ask other contributors to write in English.