Add c-lang-defconst c-anchored-cpp-prefix
This is the cause of the error that occurred in Emacs snapshots (later
25) and 26 development versions. We must confess that we did not know
how to debug Emacs Lisp and CC Mode.

PHP Mode inherits the properties of java-mode, but Java has no
preprocessor, and PHP defined `c-opt-cpp-prefix` for PHP tags.
Therefore, by properly defining `c-anchored-cpp-prefix`, errors are
prevented from occurring.

This error was c-macro-is-genuine-p called from c-beginning-of-macro,
but this problem is obscured by the huge `condition-case` of
`c-syntactic-re-search-forward`.