prevent recursion in php-syntax-propertize-extend-region
Similar to problem #277, I am receiving the error

   (error "parse-sexp-propertize-function did not move syntax-propertize--done")

due to recursion depth being exceeded.  I am using the current git branch of
emacs.  Seems that re-search function can now trigger
syntax-propertize-extend-region-functions to be called again, which in turn
triggers the same call to php-syntax-propertize-extend-region.  This stops
recursion when parameters already being processed are passed to
php-syntax-propertize-extend-region.  This fixes the error and seems to improve
handling of large files, though I have not yet profiled it.