Remove php-mode-extra-constants variable and functions
In WebMode, the `web-mode-extra-php-constants` variable has already
been removed, so it doesn't actually work. Also, the "not-extra"
constant was removed from PHP Mode in 2014, and it was out of the
role.

This function is deleted before rewriting the constant display
pattern. Add If you want to highlight keywords, we provide tips on how
to add keywords by your own hands.